+++ app/src/main/java/com/scipath/becomeaking/view/activity/GameActivity.java
@@ -99,7 +99,7 @@ public class GameActivity extends AppCompatActivity {
             CategoriesFragment newFragment = new CategoriesFragment();
             Bundle args = new Bundle();
             args.putInt("fromIndex", 10);
-            args.putInt("toIndex", 11);
+            args.putInt("toIndex", 12);
             newFragment.setArguments(args);
             switchMenuButton(buttonJob, newFragment);
         });